pkp/pkp-lib#11576 Add extensibility
